+++ packages/gamut-labs/src/experimental/AppHeader/AppHeaderElements/SharedStyles.tsx
@@ -1,0 +1,18 @@
+import { colors } from '@codecademy/gamut-styles';
+import { css } from '@emotion/react';
+
+export const focusStyles = css`
+  &:focus {
+    outline: none;
+  }
+  &:focus-visible {
+    outline: 0.3rem auto ${colors.hyper};
+  }
+`;
+
+export const hoverStyles = css`
+  &:hover {
+    color: ${colors.hyper};
+    text-decoration: none;
+  }
+`;